Testing the initial state of the task slice function
