+++ src/task.test.js
@@ -1,0 +1,9 @@
+import reducer, { tasksSlice } from './features/task/taskSlice';
+
+test('should return initial state', () => {
+  expect(
+    reducer(undefined, {
+      type: undefined,
+    })
+  ).toEqual(tasksSlice.getInitialState());
+})
